Update README.md
Made in Scratch 1.4 and in  48 hours.

This game is based on reality. God said that many people are on the road to destruction, they're going with the flow, that ultimately leads to the lake of fire. We each have to turn away from our wicked sins. 

Jump from rock to rock to save people, and don't go to the top of the screen, which is a drop off.